headings for codex entries, bio scan navigation

+++ app/src/main/java/elite/intel/gameapi/journal/BioSampleDistanceCalculator.java
@@ -11,7 +11,7 @@ public class BioSampleDistanceCalculator {
 
     public static boolean isFarEnoughFromSample(String genus, String species, double scanLat, double scanLong, double positionLat, double positionLong, double planetRadius) {
         BioForms.BioDetails details = BioForms.getDetails(genus, species);      // primary data
-        Map<String, Double> genusDistanceMap = BioScanDistances.GENUS_TO_CCR;   //fall back
+        Map<String, Integer> genusDistanceMap = BioScanDistances.GENUS_TO_CCR;   //fall back
         double requiredDistance = details == null ? genusDistanceMap.get(genus) : details.colonyRange();
         double distanceFromSample = DistanceCalculator.calculateSurfaceDistance(scanLat, scanLong, positionLat, positionLong, planetRadius);
         return distanceFromSample > requiredDistance;
